// WADNR.Web/src/app/shared/generated/model/invoice-payment-request-grid-row.ts
export class InvoicePaymentRequestGridRow { 
    InvoicePaymentRequestID?: number;
    ProjectID?: number;
    InvoicePaymentRequestDate?: string;
    VendorID?: number | null;
    VendorName?: string | null;
    VendorAddress?: string | null;
    VendorStatewideVendorNumber?: string | null;
    PreparedByPersonID?: number | null;
    PreparedByPersonFullName?: string | null;
    PreparedByPersonPhone?: string | null;
    PurchaseAuthority?: string | null;
    PurchaseAuthorityIsLandownerCostShareAgreement?: boolean;
    Duns?: string | null;
    Notes?: string | null;
    InvoiceCount?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}